Chore: Fix the build to account for moved references
